# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Prep Pal API
  version: 0.1.0
  description: Initial stable API contract for Prep Pal integration.
servers:
  - url: /api/v1
security:
  - cookieSession: []
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: prep_pal_session
...
paths:
...
  /workspaces/select:
    post:
      summary: Select active workspace.
      responses:
        "200":
          description: Active workspace selected.
        "403":
          description: User is not a member of workspace.
...